backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.0
httpx>=0.27.1
mcp>=1.5.0
google-generativeai==0.7.2
google-adk==0.3.0
pydantic==2.9.2